+++ src/parser/lex.rs
@@ -160,7 +160,7 @@ pub(crate) const USE_CASE_BODY_STARTERS: &[&[u8]] = &[
     b"then",
 ];
 
-pub(crate) const CONSTRAINT_DEF_BODY_STARTERS: &[&[u8]] = &[b"doc", b"in", b"out", b"inout"];
+pub(crate) const CONSTRAINT_DEF_BODY_STARTERS: &[&[u8]] = &[b"doc", b"@", b"in", b"out", b"inout"];
 
 pub(crate) const CALC_DEF_BODY_STARTERS: &[&[u8]] = &[b"doc", b"in", b"out", b"inout", b"return"];